move mocked task to module utils
